Tests/LibWeb/Text/expected/hit_testing/stacking-context-inside-scrollable-box.txt

  Establishes a stacking context.     <DIV id="scroll-space-filler" >
